Updates to sensitivity and visualization tests
