--- src/model_analysis/__init__.py
@@ -1,4 +1,0 @@
-from program_analysis.for2py.genPGM import (
-    create_grfn_dict,
-    get_asts_from_files,
-)
