+++ src/model_analysis/sensitivity.py
@@ -4,7 +4,7 @@ from SALib.sample import saltelli
 from SALib.analyze import sobol
 import numpy as np
 import json
-from model_analysis.networks import ComputationalGraph
+from model_assembly.networks import GroundedFunctionNetwork
 from model_analysis.utils import timeit
 
 
@@ -182,7 +182,7 @@ class SensitivityAnalyzer(object):
             for name, vector in zip(problem["names"], vectorized_sample_list)
         }
 
-        outputs = CG.run(vectorized_input_samples)
+        outputs = CG(vectorized_input_samples)
         Y = outputs[0]
         Y = Y.reshape((Y.shape[0],))
         return Y
@@ -191,7 +191,7 @@ class SensitivityAnalyzer(object):
     def Si_from_Sobol(
         cls,
         N: int,
-        G: ComputationalGraph,
+        G: GroundedFunctionNetwork,
         B: dict,
         C: dict = None,
         V: dict = None,
@@ -204,7 +204,7 @@ class SensitivityAnalyzer(object):
         """Generates Sensitivity indices using the Sobol method
         Args:
             N: The number of samples to analyze when generating Si
-            G: The ComputationalGraph to analyze
+            G: The GroundedFunctionNetwork to analyze
             B: A dictionary of bound information for the inputs of G
             C: A dictionary of cover values for use when G is a FIB
             V: A dictionary of GrFN input variable types
@@ -246,7 +246,7 @@ class SensitivityAnalyzer(object):
     def Si_from_FAST(
         cls,
         N: int,
-        G: ComputationalGraph,
+        G: GroundedFunctionNetwork,
         B: dict,
         C: dict = None,
         V: dict = None,
@@ -284,7 +284,7 @@ class SensitivityAnalyzer(object):
     def Si_from_RBD_FAST(
         cls,
         N: int,
-        G: ComputationalGraph,
+        G: GroundedFunctionNetwork,
         B: dict,
         C: dict = None,
         V: dict = None,
