+++ tests/model_analysis/test_visualization.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 import pandas as pd
 import pytest
-from model_analysis.networks import GroundedFunctionNetwork as GrFN
+from model_assembly.networks import GroundedFunctionNetwork
 from model_analysis.sensitivity import SensitivityAnalyzer
 from model_analysis.visualization import SensitivityVisualizer
 
@@ -11,7 +11,9 @@ from model_analysis.visualization import SensitivityVisualizer
 def sensitivity_visualizer():
 
     N = [10, 100, 1000, 10000]
-    tG = GrFN.from_fortran_file("tests/data/program_analysis/PETPT.for")
+    tG = GroundedFunctionNetwork.from_json(
+        "tests/data/model_analysis/PT_GrFN.json"
+    )
     var_bounds = {
         "tmax": [-30.0, 60.0],
         "tmin": [-30.0, 60.0],
